refactor: generic superclass for all message publishers
