+++ transport/base_publisher.py
@@ -1,0 +1,7 @@
+# Abstract base class for message publishers.
+
+class BasePublisher:
+
+    def publish(self):
+        raise NotImplementedError
+
